+++ changelog.md
@@ -2,6 +2,11 @@
 
 ///start typing below this text
 
+## May 24 - Adekunle's Changes
+
+- Added h2 anchor styles in `site/app/styles`
+- Added h2 anchor functionality to `site/app/scripts/app.js`
+
 ## April 7 - Adekunle's Changes
 
 - Added training as a menu item